requirements.txt
websocket-client
protobuf